Test DTD parser get_root()
Fix DTDParser.get_root() to check for cycles
and lack of a root element.
Add test for DTD without root
Add test for XML generator for nested child elements
