+++ data/dtd/no_root.dtd
@@ -1,0 +1,2 @@
+<!ELEMENT A (B)>
+<!ELEMENT B (A)>
